+++ src/main/java/school/faang/user_service/service/recommendation/request/validator/Validator.java
@@ -1,0 +1,5 @@
+package school.faang.user_service.service.recommendation.request.validator;
+
+public interface Validator<T> {
+    boolean validate(final T data);
+}
