--- src/main/java/school/faang/user_service/validator/Validator.java
@@ -1,5 +1,0 @@
-package school.faang.user_service.validator;
-
-public interface Validator<T> {
-    boolean validate(final T data);
-}
